+++ src/N98/Magento/Command/System/Cron/ListCommand.php
@@ -2,6 +2,8 @@
 
 namespace N98\Magento\Command\System\Cron;
 
+use Magento\Framework\App\Area;
+use Magento\Framework\App\AreaList;
 use N98\Util\Console\Helper\Table\Renderer\RendererFactory;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -30,9 +32,20 @@ class ListCommand extends AbstractCronCommand
      * @param \Symfony\Component\Console\Input\InputInterface $input
      * @param \Symfony\Component\Console\Output\OutputInterface $output
      * @return int|void
+     * @throws \Magento\Framework\Exception\LocalizedException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $this->state->setAreaCode(Area::AREA_CRONTAB);
+
+        $objectManager = $this->getObjectManager();
+
+        /** @var AreaList $areaList */
+        $areaList = $objectManager->get(AreaList::class);
+        $areaList->getArea(Area::AREA_CRONTAB)
+            ->load(Area::PART_CONFIG)
+            ->load(Area::PART_TRANSLATE);
+
         if ($input->getOption('format') === null) {
             $this->writeSection($output, 'Cronjob List');
         }